numdoorfix

+++ TrafficSim/src/com/company/Car.java
@@ -2,9 +2,9 @@ package com.company;
 
 final public class Car extends Vehicle
 {
-    private Double numberDoors;
+    private int numberDoors;
 
-    public Car(MovementStatus ms, String color, Double size, Double weight, Double maxSpeed, Reputation rep, DamageStatus ds, Double nd)
+    public Car(MovementStatus ms, String color, Double size, Double weight, Double maxSpeed, Reputation rep, DamageStatus ds, int nd)
     {
         super(ms, color, size, weight, maxSpeed, rep, ds);
         numberDoors = nd;
